logic client html
